refactor(identity): use one-time random key for registration (DIP-0011)
