+++ src/types/AssetLock.ts
@@ -6,18 +6,10 @@ export interface BuildAssetLockFromFundingTxOptions {
   network: string
   /** Hash of the funding transaction whose output is being spent. */
   assetLockFundingTxid: string
-  /** Address that owns the funding UTXO (input being spent). */
+  /** One-time funding address (input UTXO + asset lock credit output). */
   assetLockFundingAddress: string
-  /** WIF of the key that controls assetLockFundingAddress (signs the input). */
+  /** WIF of the one-time key that controls assetLockFundingAddress. */
   assetLockFundingPrivateKeyWif: string
-  /**
-   * Address that receives the asset lock credit output (in the payload).
-   * Per DIP-0011 this address must own the key that signs the
-   * IdentityCreateTransition. Decoupled from assetLockFundingAddress so the
-   * registration key can be derived stateless from the seed at the on-chain
-   * free identityIndex.
-   */
-  creditOutputAddress: string
   /** Optional: index of the funding output in the funding transaction. */
   outputIndex?: number
 }
